svn:revision:37287
svn:branch:STABLE
svn:account:starseeker

+++ src/libtclcad/tkImgFmtPIX.c
@@ -1,7 +1,7 @@
 /*                   T K I M G F M T P I X . C
  * BRL-CAD
  *
- * Copyright (c) 1994-2009 United States Government as represented by
+ * Copyright (c) 1994-2010 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -197,7 +197,7 @@ FileReadPIX(Tcl_Interp *interp, Tcl_Channel chan, const char *fileName, Tcl_Obj 
     /* Coordinates of top-left pixel to be used
      * in image being read. */
 {
-    size_t fileWidth, fileHeight;
+    unsigned long int fileWidth, fileHeight;
     int nBytes, h, count;
     unsigned char *pixelPtr;
     Tk_PhotoImageBlock block;
